Create a separate component for position details
